# swagger.yaml
openapi: 3.0.0
info:
  title: Find Your Vibe API
  description: |
    API documentation for the Find Your Vibe event matching application.
    
    The matching algorithm now considers sexual orientation and relationship preferences as key factors in creating compatible matches.
  version: 1.1.0
servers:
  - url: http://localhost:5000/api
    description: Local development server
components:
  schemas:
    Event:
      type: object
      properties:
        eventCode:
          type: string
          description: Unique 6-character event code
        hostName:
          type: string
          description: Name of the event host
        questions:
          type: array
          items:
            type: object
            properties:
              text:
                type: string
              options:
                type: array
                items:
                  type: string
        countdownDuration:
          type: integer
          description: Event countdown duration in seconds
        participantCount:
          type: integer
          description: Number of participants in the event
        startTime:
          type: string
          format: date-time
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
paths:
  /events:
    post:
      summary: Create a new event
      description: Create a new event with host information and optional custom questions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - hostName
              properties:
                hostName:
                  type: string
                questions:
                  type: array
                  items:
                    type: object
                    properties:
                      text:
                        type: string
                      options:
                        type: array
                        items:
                          type: string
                countdownDuration:
                  type: integer
                  description: Duration in seconds (default 3600)
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  eventCode:
                    type: string
                  event:
                    $ref: '#/components/schemas/Event'
        '400':
          description: Invalid request parameters
        '500':
          description: Server error
  /events/{eventCode}:
    get:
      summary: Get event details
      parameters:
        - name: eventCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Event not found
  /events/{eventCode}/start:
    post:
      summary: Start event countdown
      parameters:
        - name: eventCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event started successfully
        '404':
          description: Event not found
...
  /events/{eventCode}/qrcode:
    get:
      summary: Generate a QR code for joining an event
      description: Creates a QR code image (as a data URL) containing a link that participants can scan to join the event
      parameters:
        - name: eventCode
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: QR code generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  eventCode:
                    type: string
                    description: The event code
                  qrCode:
                    type: string
                    description: Data URL containing the QR code image
                  joinUrl:
                    type: string
                    description: The URL encoded in the QR code
        '404':
          description: Event not found
        '500':
          description: Server error 